+++ graph/actions.go
@@ -71,6 +71,23 @@ func (di *SousGraph) GetManifestSet(dff config.DeployFilterFlags, up *restful.Up
 	}, nil
 }
 
+// GetPlumbingNormalizeGDM returns an update Action.
+func (di *SousGraph) GetPlumbingNormalizeGDM() (actions.Action, error) {
+	scoop := struct {
+		LS     LogSink
+		User   sous.User
+		Config LocalSousConfig
+	}{}
+	if err := di.Inject(&scoop); err != nil {
+		return nil, err
+	}
+	return &actions.PlumbNormalizeGDM{
+		User:          scoop.User,
+		StateLocation: scoop.Config.StateLocation,
+		Log:           scoop.LS.LogSink.Child("plumbing-normalize-gdm"),
+	}, nil
+}
+
 // GetUpdate returns an update Action.
 func (di *SousGraph) GetUpdate(dff config.DeployFilterFlags, otpl config.OTPLFlags) (actions.Action, error) {
 	di.guardedAdd("DeployFilterFlags", &dff)